refactor: 문의 관련 파일 분리 / 주특기 선택 리팩터링
문의관련된 컨트롤러와 서비스를 분리하였으며, 주특기 선택에 대한 메서드를 리팩터링하였습니다.
---
유저 파트 확정에 대한 기능을 관리자로 넘겨야 하지 않을까에 대한 의문이 생겨 이슈 생성
