+++ src/main/java/org/supercoding/supertime/user/util/UserValidation.java
@@ -10,6 +10,8 @@ import org.supercoding.supertime.golbal.web.advice.CustomNoSuchElementException;
 import org.supercoding.supertime.golbal.web.advice.CustomNotFoundException;
 import org.supercoding.supertime.inquiry.repository.InquiryRepository;
 import org.supercoding.supertime.inquiry.web.entity.InquiryEntity;
+import org.supercoding.supertime.semester.repository.SemesterRepository;
+import org.supercoding.supertime.semester.web.entity.SemesterEntity;
 import org.supercoding.supertime.user.repository.UserRepository;
 import org.supercoding.supertime.user.web.entity.user.UserEntity;
 
@@ -21,6 +23,7 @@ public class UserValidation {
 
     private final UserRepository userRepository;
     private final InquiryRepository inquiryRepository;
+    private final SemesterRepository semesterRepository;
 
     public UserEntity validateExistUser(String username) {
         return userRepository.findByUserId(username)
@@ -47,5 +50,10 @@ public class UserValidation {
     public InquiryEntity validateExistInquiry(Long inquiryCid) {
         return inquiryRepository.findById(inquiryCid)
                 .orElseThrow(()-> new CustomNotFoundException("해당 문의가 존재하지 않습니다."));
+    }
+
+    public SemesterEntity validateExistSemester(Long semesterCid) {
+        return semesterRepository.findById(semesterCid)
+                .orElseThrow(()-> new CustomNotFoundException("기수가 존재하지 않습니다."));
     }
 }
